fix(burgs): allocate cells.burg as Uint32Array in heightmap editor and .map load
Both sites truncated burg ids >65535 on 100K-burg maps. Guard test now
also catches Uint16Array.from, not just new Uint16Array.

+++ src/controllers/heightmap-editor.ts
@@ -637,7 +637,7 @@ function restoreRiskedData(): void {
   pack.cells.pop = new Float32Array(n);
   pack.cells.routes = {};
   pack.cells.s = new Uint16Array(n);
-  pack.cells.burg = new Uint16Array(n);
+  pack.cells.burg = new Uint32Array(n);
   pack.cells.state = new Uint16Array(n);
   pack.cells.province = new Uint16Array(n);
   pack.cells.culture = new Uint16Array(n);
